[iOS] refactor: UIView extension 코드 병합
Issue #21
