--- iOS/AirbnbProject/AirbnbProject/Util/extensionUIView.swift
@@ -1,20 +1,0 @@
-//
-//  extensionUIView.swift
-//  AirbnbProject
-//
-//  Created by 임승혁 on 2020/05/21.
-//  Copyright © 2020 dev-Lena. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-extension UIView {
-    func setupXIB(nibName: String) {
-        let bundle = Bundle(for: type(of: self))
-        let nib = UINib(nibName: nibName, bundle: bundle)
-        let view = nib.instantiate(withOwner: self, options: nil).first as! UIView
-        addSubview(view)
-        view.frame = self.bounds
-    }
-}
